Refactory :scss 줄넘김 & variable 변수 설정 & 댓글 삭제 수정
