+++ src/Pages/Juan/Login/Login.scss
@@ -1,16 +1,18 @@
+@import '../../../styles/variable.scss';
+
 .login {
   display: flex;
   justify-content: center;
   align-items: center;
   height: 100vh;
-  background-color: #fafafa;
+  background-color: $background-color;
 
   .container {
     display: flex;
     flex-direction: column;
     align-items: center;
     background-color: white;
-    border: 1px solid rgba(0, 0, 0, 0.1);
+    border: $boder-light;
     padding: 3rem;
     min-height: 80vh;
   }
@@ -26,9 +28,9 @@
     font-size: 15px;
     margin-bottom: 20px;
     padding: 1rem;
-    border: 1px solid rgba(0, 0, 0, 0.1);
+    border: $boder-light;
     border-radius: 5px;
-    background-color: #fafafa;
+    background-color: $background-color;
   }
 
   .loginBtn {
