+++ src/Pages/Juan/Main/components/Aside.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import { USER_DATA } from '../Main';
+import Story from './Story';
+import RecommendUser from './RecommendUser';
 
 function Aside() {
   return (
-    <section className="mainRight">
+    <section className="aside">
       <div className="rightItem1">
         <div className="userProfileImgCircle user_img" />
         <div className="user_tag">
@@ -12,17 +13,17 @@ function Aside() {
         </div>
       </div>
 
-      <section className="rightItem2">
+      <section className="aside_story">
         <div className="right_item_header">
           <p className="right_item_title">스토리</p>
           <p className="right_item_all">모두보기</p>
         </div>
         <div className="user_story_container">
-          <UserStoryContainer />
+          <Story />
         </div>
       </section>
 
-      <div className="rightItem3">
+      <div className="aside_recommend">
         <div className="right_item_header">
           <p className="right_item_title">회원님을 위한 추천</p>
           <p className="right_item_all">모두보기</p>
@@ -35,42 +36,4 @@ function Aside() {
   );
 }
 
-function UserStoryContainer() {
-  return USER_DATA.map(user => {
-    return (
-      <li className="user_container" key={user.id}>
-        <img
-          src={user.profile_url}
-          alt="유저 이미지"
-          className="userProfileImgCircle article_profile onStory"
-        />
-        <div className="user_tag">
-          <p className="userId">{user.user_id}</p>
-          <p className="article_profile_content">{user.time}</p>
-        </div>
-      </li>
-    );
-  });
-}
-
-function RecommendUser() {
-  return USER_DATA.map(user => {
-    return (
-      <div className="user_container" key={user.id}>
-        <img
-          src={user.profile_url}
-          alt="유저 이미지"
-          className="userProfileImgCircle"
-        />
-        <div className="user_tag">
-          <p className="userId">{user.user_id}</p>
-          <p className="article_profile_follow">
-            _legend_a님 외 2명이 팔로우하고 있습니다
-          </p>
-        </div>
-        <p className="follow_btn">팔로우</p>
-      </div>
-    );
-  });
-}
 export default Aside;
