+++ src/index.js
@@ -3,6 +3,6 @@ import ReactDOM from 'react-dom/client';
 import Router from './Router';
 import './styles/reset.scss';
 import './styles/common.scss';
-
+import './styles/variable.scss';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(<Router />);
