+++ src/styles/variable.scss
@@ -1,0 +1,4 @@
+$background-color: #fafafa;
+$boder-light: 1px solid rgba(0, 0, 0, 0.1);
+$gray: rgba(0, 0, 0, 0.5);
+$bg-size: center center cover no-repeat;
